% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_ash.R
\name{ash1}
\alias{ash1}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
ash1(bins, m = 5, kopt = c(2, 2), quiet = TRUE)
}
\arguments{
\item{bins}{
(input list) \code{$nc}=integer vector of bin counts and \code{$ab}=bin interval
}

\item{m}{
(input) optional integer smoothing parameter;  default=5.
}

\item{kopt}{
(input) vector of length 2 specifying the kernel, which
is proportional to \eqn{( 1 - abs(i/m)^kopt(1) )i^kopt(2)};
(2,2)=biweight (default);  (0,0)=uniform;  (1,0)=triangle;
(2,1)=Epanechnikov;  (2,3)=triweight.
}
}
\value{
returns structure suitable for input to \code{plot}
dd
\item{x=t}{
vector of bin center locations
}
\item{y=f}{
vector of ash estimates
}
\item{ier}{
0=normal exit; 1=estimate nonzero outside interval ab
}
}
\description{
ash::ash1 modified to use message() instead of regular text output
}
\seealso{
\code{\link{bin1}}
}
\keyword{internal}
