% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dithering.R
\name{Dithering}
\alias{Dithering}
\title{Apply dithering to integer values}
\usage{
Dithering(x)
}
\arguments{
\item{x}{integer vector or matrix.}
}
\value{
\code{Dithering} returns a numeric vector or matrix.
}
\description{
The \code{Dithering} function applies a triangular dithering filter
which should eliminate the presence of identical non-zero values
in the input vector or matrix, whithout distorting the original
distribution of values.
}
