% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_other_functions.R
\name{FiniteValues}
\alias{FiniteValues}
\title{Localize safe numeric observations (i.e. not NA, NaN, Inf)}
\usage{
FiniteValues(x)
}
\arguments{
\item{x}{numeric vector or matrix.}
}
\value{
\code{FiniteValues} returns a logical vector.
}
\description{
Localize safe numeric observations (i.e. not NA, NaN, Inf)
}
\seealso{
\link{is.finite}
}
\keyword{internal}
