% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_knn_stats.R
\name{knn_mean}
\alias{knn_mean}
\title{Multivariate local means}
\usage{
knn_mean(x, k, xref = NULL, nn_alg = "kd_tree")
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions (query data).}

\item{k}{number of nearest neighbors which has equivalent effects as the
usual bandwidth or smoothing parameters (larger k => smoother results).}

\item{xref}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions (optional reference data).}

\item{nn_alg}{nearest neighbor searching algorithm (default = "kd_tree").
See \link{get.knn} for further explanations.}
}
\value{
\code{knn_mean} returns a matrix of knn centroids.
}
\description{
compute k-nearest neighbor (knn) centroids (i.e. mean vectors).
}
\seealso{
\link{knn_density}
}
