% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_knn_stats.R
\name{knn_values}
\alias{knn_values}
\title{Extract multivariate local values based on a knn index matrix}
\usage{
knn_values(v, i)
}
\arguments{
\item{v}{numeric vector.}

\item{i}{precomputed matrix of nearest neighbor indexes.}
}
\value{
\code{knn_values} returns a matrix.
}
\description{
Extract multivariate local values based on a knn index matrix
}
\keyword{internal}
