% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_matrixes.R
\name{m2v}
\alias{m2v}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
m2v(i, j, nrow)
}
\description{
Convert matrix indices into vector indices such that:

x <- m2v(i, j) => M[i, j] = M[x] for each pair of indices
}
\keyword{internal}
