% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchClusters.R
\name{MatchClusters}
\alias{MatchClusters}
\title{Optimal match between two different cluster/class assignments}
\usage{
MatchClusters(x, r, n = NULL)
}
\arguments{
\item{x}{integer vector representing a set of cluster/class memberships.}

\item{r}{integer vector representing a set of alternative cluster/class memberships
(e.g. reference or ground truth assignments).}

\item{n}{total number of classes, inferred from argument \code{x} when not provided.}
}
\value{
\code{MatchClusters} returns a matrix with the following columns:
\item{x2r}{x to r cluster/class translation}
\item{r2x}{r to x cluster/class translation}
}
\description{
Optimal match between two different cluster/class assignments
}
