% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plots.R
\name{PlotQuickShift}
\alias{PlotQuickShift}
\title{PlotQuickShift}
\usage{
PlotQuickShift(x, g, new = TRUE, length = 0.05, col = rgb(0, 0, 0,
  0.2), ...)
}
\arguments{
\item{x}{matrix}

\item{g}{graph}

\item{...}{}
}
\description{
PlotQuickShift
}
\keyword{internal}
