% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickShift.R
\name{QuickShift}
\alias{QuickShift}
\title{Hierarchical clustering based on the estimated density gradient}
\usage{
QuickShift(x, n, d = NULL, k = NULL, q = 2, ...)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions.}

\item{n}{desired number of clusters.}

\item{d}{numeric vector representing a density estimation at each observation.}

\item{k}{number of nearest neighbors for density estimation (used only when argument
\code{d} is not provided).}

\item{q}{number of nearest neighbors used to find increasing densities during the
QuickShift iterations. The minimum value is 2 (default).}

\item{...}{optional arguments passed to the \link{QuickShiftAlgorithm} or
\link{QuickShiftClusters} functions.}
}
\value{
This function returns a list with the following elements:
\item{membership}{
  vector of integers in [1, \code{nclust}] indicating to which cluster each
  observation belongs.
}
\item{csizes}{number of observations in each cluster.}
\item{nclust}{number of clusters.}
And optionally:
\item{graph}{the QuickShift graph (see \link{QuickShiftAlgorithm}).}
}
\description{
Hierarchical clustering based on the estimated density gradient
}
\references{
Vedaldi A., Soatto S. (2008) Quick Shift and Kernel Methods for Mode Seeking.
In: Forsyth D., Torr P., Zisserman A. (eds) Computer Vision – ECCV 2008.
ECCV 2008. Lecture Notes in Computer Science, vol 5305.
Springer, Berlin, Heidelberg
\url{http://dx.doi.org/10.1007/978-3-540-88693-8_52}
}
\seealso{
\link{QuickShiftAlgorithm},
  \link{QuickShiftClusters},
  \link{knn_density}
}
