% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_quickshift.R
\name{QuickShiftAlgorithmAlt}
\alias{QuickShiftAlgorithmAlt}
\title{QuickShift algorithm (Vedaldi & Soatto, 2008) - alternative implementation}
\usage{
QuickShiftAlgorithmAlt(x, d, k = 3, plot = FALSE, alg = 0)
}
\description{
QuickShift algorithm (Vedaldi & Soatto, 2008) - alternative implementation
}
\keyword{internal}
