% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_quickshift.R
\name{QuickShiftClusters}
\alias{QuickShiftClusters}
\title{Split a QuickShift graph into clusters}
\usage{
QuickShiftClusters(g, n)
}
\arguments{
\item{g}{QuickShift graph resulting from the \link{QuickShiftAlgorithm} function.}

\item{n}{desired number of clusters.}
}
\value{
\code{QuickShift} and \code{QuickShiftClusters} return a list
with the same following elements:
\item{membership}{
  vector of integers in [1, \code{n}] indicating to which cluster each
  observation belongs.
}
\item{sizes}{number of observations in each cluster.}
}
\description{
Split a QuickShift graph into clusters
}
\seealso{
\link{QuickShiftAlgorithm},
  \link{QuickShift}
}
\keyword{internal}
