% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utilities.R
\name{RankScore}
\alias{RankScore}
\title{Rescale x non-linearly into ]0, 1[}
\usage{
RankScore(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
\code{RankScore} returns a numeric vector of rank scores.
}
\description{
rescale values non-linearly to the unit interval using rank scores \eqn{q}
given by \eqn{q = (rank(x) - 0.5) / N} where \eqn{N} = length(x).
}
\seealso{
\link{S01},
  \link{SX2Y}
}
\keyword{internal}
