% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utilities.R
\name{S01}
\alias{S01}
\title{Rescale x linearly into [0, 1]}
\usage{
S01(x)
}
\arguments{
\item{x}{numeric vector or matrix.}
}
\value{
\code{S01} returns x linearly rescaled such that range(x) = [0, 1].
}
\description{
rescale values linearly to the unit interval.
}
\seealso{
\link{RankScore},
  \link{SX2Y}
}
\keyword{internal}
