% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_other_functions.R
\name{v2m}
\alias{v2m}
\title{** RESERVED FOR INTERNAL USE **}
\usage{
v2m(x, nrow)
}
\description{
Convert vector indexes into matrix indexes such that:

ij <- v2m(x) => M[x] = M[ij[, 1], ij[, 2]] for each pair of indexes
}
\keyword{internal}
