% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_knn_stats.R
\name{knn_density}
\alias{knn_density}
\title{knn density estimator}
\usage{
knn_density(x, k, data = NULL, i = NULL, d = NULL, smoothing = TRUE,
  sum2one = TRUE)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions (query).}

\item{k}{number of nearest neighbors, which corresponds to the smoothing parameter
of estimated densities (larger k values = smoother).}

\item{i}{precomputed matrix of nearest neighbor indices (optional).}

\item{d}{precomputed matrix of nearest neighbor distances (optional).}

\item{smoothing}{perfom a local average smoothing of
the estimated density for \link{knn_density} (default = TRUE)
or the local variance for \link{knn_musigma2} (default = FALSE).}

\item{y}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions.}
}
\value{
\code{knn_density} returns a numeric vector.
}
\description{
k-nearest neighbor (knn) estimator of the density \deqn{P(Xi) ~ k / N Vi}
where \eqn{N} is the number of observations and \eqn{Vi} the volume of a
sphere of radius equal to the distance between \eqn{Xi} and its k-nearest
neighbor.
}
\seealso{
\link{knn_musigma2},
  \link{knn_mean}
}
\keyword{internal}
