% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_knn_stats.R
\name{knn_musigma2}
\alias{knn_musigma2}
\title{Local mean and variance}
\usage{
knn_musigma2(x, k, xref = NULL, smoothing = FALSE,
  nn_alg = "kd_tree")
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions (query data).}

\item{k}{number of nearest neighbors which has equivalent effects as the
usual bandwidth or smoothing parameters (larger k => smoother results).}

\item{xref}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions (optional reference data).}

\item{smoothing}{logical value activating an additional local average smoothing of
the estimated variances (default = FALSE, inactive).}

\item{nn_alg}{nearest neighbor searching algorithm (default = "kd_tree").
See \link{get.knn} for further explanations.}
}
\value{
\code{knn_musigma2} returns a list with the following elements:
\item{mu}{knn centroids}
\item{sigma2}{knn variances (squared standard deviations)}
}
\description{
k-nearest neighbor (knn) centroid and associated local standard deviation.
}
\seealso{
\link{knn_density},
  \link{knn_mean}
}
\keyword{internal}
