% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchClusters.R
\name{MatchClusters}
\alias{MatchClusters}
\title{Optimal match between two different cluster/class assignments}
\usage{
MatchClusters(x, r)
}
\arguments{
\item{x}{integer vector representing a set of cluster/class memberships.}

\item{r}{integer vector representing an alternative set of cluster/class memberships
(e.g. reference or ground truth assignments).}
}
\value{
\code{MatchClusters} returns an integer matrix with the following columns:
\item{x2r}{translation from x to r cluster/class identifiers}
\item{r2x}{translation from r to x cluster/class identifiers}
For instance, noting \code{m} such translation matrix, then
\item{m[3, "x2r"]}{
  is the best matching cluster identifier \code{r} for the cluster 3
  in classification \code{x}
}
\item{m[5, "r2x"]}{
  is the best matching cluster identifier \code{x} for the cluster 5
  in classification \code{r}
}
}
\description{
Optimal match between two different cluster/class assignments
}
