% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotQuickShift.R
\name{PlotQuickShift}
\alias{PlotQuickShift}
\title{Plot a QuickShift graph}
\usage{
PlotQuickShift(x, g, new = TRUE, length = 0.05, col = grey(0, 0.2),
  ...)
}
\arguments{
\item{x}{numeric matrix representing 2D coordinates of the data points.}

\item{g}{a QuickShift graph resulting from the \link{QuickShiftAlgorithm} function.}

\item{new}{logical value whether to start a new plot (default = TRUE, yes).}

\item{length}{numeric value controlling arrow lengths (default = 0.05).}

\item{col}{arrow colors (\code{grey(0, 0.2)} by default).}

\item{...}{optional arguments passed to the \link{arrows} function.}
}
\description{
Plot a QuickShift graph
}
