% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_quickshift.R
\name{QuickShiftAlgorithm}
\alias{QuickShiftAlgorithm}
\title{QuickShift algorithm (Vedaldi & Soatto, 2008)}
\usage{
QuickShiftAlgorithm(x, d, k = 2, plot = FALSE, simplified = TRUE)
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions.}

\item{d}{numeric vector representing a density estimation at each observation.}

\item{k}{number of nearest neighbors used to find increasing densities.
The minimum value of is 2 (default).}

\item{plot}{logical value allowing to visualize the contruction of the QuickShift graph.}

\item{simplified}{logical value enabling to use a specific and faster implementation
when argument \code{k} is equal to 2 (default = TRUE, yes).}
}
\value{
\code{QuickShiftAlgorithm} returns a graph object (see \link{igraph} package).
}
\description{
QuickShift algorithm (Vedaldi & Soatto, 2008)
}
\references{
Vedaldi A., Soatto S. (2008) Quick Shift and Kernel Methods for Mode Seeking.
In: Forsyth D., Torr P., Zisserman A. (eds) Computer Vision – ECCV 2008.
ECCV 2008. Lecture Notes in Computer Science, vol 5305.
Springer, Berlin, Heidelberg
\url{http://dx.doi.org/10.1007/978-3-540-88693-8_52}
}
\seealso{
\link{QuickShiftClusters},
  \link{QuickShift}
}
\keyword{internal}
