% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickShiftAlgorithm.R
\name{QuickShiftAlgorithm}
\alias{QuickShiftAlgorithm}
\title{QuickShift algorithm (Vedaldi & Soatto, 2008)}
\usage{
QuickShiftAlgorithm(x, d, plot = FALSE, nn_alg = "kd_tree")
}
\arguments{
\item{x}{numeric matrix representing multivariate data where rows = observations
and columns = measurement conditions.}

\item{d}{numeric vector representing a density estimation at each observation.}

\item{plot}{logical value allowing to visualize the contruction of the QuickShift graph.}

\item{nn_alg}{nearest neighbor searching algorithm (default = "kd_tree").
See \link{get.knn} for further explanations.}
}
\value{
\code{QuickShiftAlgorithm} returns a graph object
(see \link{igraph} package). The \link{QuickShiftClusters} function
can extract clusters from such graph.
}
\description{
Implementation of the QuickShift algorithm using accelerated nearest
neighbor queries from the FNN package.
}
\references{
Vedaldi A., Soatto S. (2008) Quick Shift and Kernel Methods for Mode Seeking.
In: Forsyth D., Torr P., Zisserman A. (eds) Computer Vision – ECCV 2008.
ECCV 2008. Lecture Notes in Computer Science, vol 5305.
Springer, Berlin, Heidelberg
\url{http://dx.doi.org/10.1007/978-3-540-88693-8_52}
}
\seealso{
\link{QuickShiftClusters},
  \link{QuickShift}
}
