% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickShiftClusters.R
\name{QuickShiftClusters}
\alias{QuickShiftClusters}
\title{Clustering from a QuickShift graph}
\usage{
QuickShiftClusters(g, n, graph = FALSE, id_by_pop = TRUE, decreasing = TRUE)
}
\arguments{
\item{g}{a QuickShift graph resulting from the \link{QuickShiftAlgorithm} function.}

\item{n}{desired number of clusters.}

\item{graph}{logical value indicating whether the QuickShift graph should be part of
the returned results (default = FALSE, no).}

\item{id_by_pop}{logical value controlling if cluster identifiers should follow the order
of population sizes (default = TRUE, yes).}

\item{decreasing}{logical value indicating if population sizes should be in decreasing
(default = TRUE, yes) or increasing order, when the argument
\code{id_by_pop} is TRUE.}
}
\value{
This function returns a list with the following elements:
\item{membership}{
  vector of integers in [1, \code{nclust}] indicating to which cluster each
  observation belongs.
}
\item{csizes}{number of observations in each cluster.}
\item{nclust}{number of clusters.}
And optionally:
\item{graph}{the QuickShift graph (see \link{QuickShiftAlgorithm}).}
}
\description{
Clustering from a QuickShift graph
}
\seealso{
\link{QuickShift},
  \link{MatchClusters}
}
